Refactor Renderer and use double everywhere
Refactor Renderer class to have RenderPixel and WriteOutput methods, separating colour calculation from writing to file. Also switch from float to double in all instances for greater precision.

+++ RayTracer/src/Camera.cpp
@@ -1,19 +1,19 @@
 #include "Camera.h"
 
 Camera::Camera() {
-	float aspectRatio = 16.0f / 9.0f;
-	float viewportHeight = 2.0f;
-	float viewportWidth = aspectRatio * viewportHeight;
-	float focalLength = 1.0f;
+	double aspectRatio = 16.0 / 9.0;
+	double viewportHeight = 2.0;
+	double viewportWidth = aspectRatio * viewportHeight;
+	double focalLength = 1.0;
 
-	m_Origin = glm::vec3(0.0f, 0.0f, 0.0f);
-	m_ViewportHorizontal = glm::vec3(viewportWidth, 0.0f, 0.0f);
-	m_ViewportVertical = glm::vec3(0.0f, viewportHeight, 0.0f);
-	m_ViewportLowerLeftCorner = m_Origin - m_ViewportHorizontal / 2.0f - m_ViewportVertical / 2.0f - glm::vec3(0.0f, 0.0f, focalLength);
+	m_Origin = glm::dvec3(0.0, 0.0, 0.0);
+	m_ViewportHorizontal = glm::dvec3(viewportWidth, 0.0, 0.0);
+	m_ViewportVertical = glm::dvec3(0.0, viewportHeight, 0.0);
+	m_ViewportLowerLeftCorner = m_Origin - m_ViewportHorizontal / 2.0 - m_ViewportVertical / 2.0 - glm::dvec3(0.0, 0.0, focalLength);
 }
 
 Camera::~Camera() {}
 
-Ray Camera::getRay(float pixelX, float pixelY) const {
+Ray Camera::getRay(double pixelX, double pixelY) const {
 	return Ray(m_Origin, m_ViewportLowerLeftCorner + pixelX * m_ViewportHorizontal + pixelY * m_ViewportVertical - m_Origin);
 }
